+++ packages/runtime/src/account.ts
@@ -1,6 +1,5 @@
 import { types } from "@algo-builder/web";
 import type {
-  Account,
   AssetDef,
   SSCSchemaConfig
 } from "algosdk";
@@ -10,16 +9,17 @@ import { RUNTIME_ERRORS } from "./errors/errors-list";
 import { RuntimeError } from "./errors/runtime-errors";
 import { checkAndSetASAFields } from "./lib/asa";
 import {
-  ALGORAND_ACCOUNT_MIN_BALANCE, APPLICATION_BASE_FEE, ASSET_CREATION_FEE, MAX_ALGORAND_ACCOUNT_APPS,
-  MAX_ALGORAND_ACCOUNT_ASSETS,
-  SSC_KEY_BYTE_SLICE, SSC_VALUE_BYTES, SSC_VALUE_UINT
+  ALGORAND_ACCOUNT_MIN_BALANCE, APPLICATION_BASE_FEE,
+  ASSET_CREATION_FEE, MAX_ALGORAND_ACCOUNT_APPS,
+  MAX_ALGORAND_ACCOUNT_ASSETS, SSC_KEY_BYTE_SLICE,
+  SSC_VALUE_BYTES, SSC_VALUE_UINT
 } from "./lib/constants";
 import { keyToBytes } from "./lib/parsing";
 import { assertValidSchema } from "./lib/stateful";
 import {
-  AccountStoreI,
-  AppDeploymentFlags, AppLocalStateM, AssetHoldingM, CreatedAppM, SSCAttributesM,
-  StackElem
+  AccountStoreI, AppDeploymentFlags, AppLocalStateM,
+  AssetHoldingM, CreatedAppM, RuntimeAccount,
+  SSCAttributesM, StackElem
 } from "./types";
 
 const StateMap = "key-value";
@@ -29,7 +29,7 @@ const globalStateSchema = "global-state-schema";
 const numByteSlice = 'num-byte-slice';
 
 export class AccountStore implements AccountStoreI {
-  readonly account: Account;
+  readonly account: RuntimeAccount;
   readonly address: string;
   minBalance: number; // required minimum balance for account
   assets: Map<number, AssetHoldingM>;
@@ -39,8 +39,12 @@ export class AccountStore implements AccountStoreI {
   createdApps: Map<number, SSCAttributesM>;
   createdAssets: Map<number, AssetDef>;
 
-  constructor (balance: number | bigint, account?: Account) {
-    if (account) {
+  constructor (balance: number | bigint, account?: RuntimeAccount | string) {
+    if (typeof account === 'string') {
+      this.account = generateAccount();
+      this.account.name = account;
+      this.address = this.account.addr;
+    } else if (account) {
       // set config if account is passed by user
       this.account = account;
       this.address = account.addr;